requirements.txt
accelerate==1.7.0
aiofiles==24.1.0
altair @ file:///home/conda/feedstock_root/build_artifacts/altair-split_1734244716962/work
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
async-timeout==4.0.3
attrs @ file:///home/conda/feedstock_root/build_artifacts/attrs_1741918516150/work
backoff==2.2.1
beautifulsoup4==4.13.4
blinker @ file:///home/conda/feedstock_root/build_artifacts/blinker_1731096409132/work
Brotli @ file:///Users/runner/miniforge3/conda-bld/brotli-split_1749229914644/work
cachetools @ file:///home/conda/feedstock_root/build_artifacts/cachetools_1740094013202/work
certifi @ file:///home/conda/feedstock_root/build_artifacts/certifi_1746569525376/work/certifi
cffi @ file:///Users/runner/miniforge3/conda-bld/cffi_1725560622100/work
chardet==5.2.0
charset-normalizer @ file:///home/conda/feedstock_root/build_artifacts/charset-normalizer_1746214863626/work
click @ file:///home/conda/feedstock_root/build_artifacts/click_1734858813237/work
coloredlogs==15.0.1
contourpy==1.3.0
cryptography==45.0.3
cycler==0.12.1
dataclasses-json==0.6.7
Deprecated==1.2.18
EbookLib==0.19
effdet==0.4.1
emoji==2.14.1
exceptiongroup==1.3.0
faiss-cpu==1.11.0
filelock==3.18.0
filetype==1.2.0
flatbuffers==25.2.10
fonttools==4.58.2
fsspec==2025.5.1
gitdb @ file:///home/conda/feedstock_root/build_artifacts/gitdb_1735887193964/work
GitPython @ file:///home/conda/feedstock_root/build_artifacts/gitpython_1735929639977/work
google-api-core==2.25.0
google-auth==2.40.3
google-cloud-vision==3.10.1
googleapis-common-protos==1.70.0
grpcio==1.72.1
grpcio-status==1.72.1
h11==0.16.0
h2 @ file:///home/conda/feedstock_root/build_artifacts/h2_1738578511449/work
hf-xet==1.1.3
hpack @ file:///home/conda/feedstock_root/build_artifacts/hpack_1737618293087/work
html5lib==1.1
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.32.4
humanfriendly==10.0
hyperframe @ file:///home/conda/feedstock_root/build_artifacts/hyperframe_1737618333194/work
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1733211830134/work
importlib_metadata @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_importlib-metadata_1747934053/work
importlib_resources @ file:///home/conda/feedstock_root/build_artifacts/importlib_resources_1736252299705/work
Jinja2 @ file:///home/conda/feedstock_root/build_artifacts/jinja2_1741263328855/work
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema @ file:///home/conda/feedstock_root/build_artifacts/jsonschema_1748294245630/work
jsonschema-specifications @ file:///tmp/tmpuvkyqc9y/src
kiwisolver==1.4.7
langchain==0.3.25
langchain-core==0.3.64
langchain-text-splitters==0.3.8
langdetect==1.0.9
langsmith==0.3.45
lxml==5.4.0
MarkupSafe @ file:///Users/runner/miniforge3/conda-bld/markupsafe_1733219663979/work
marshmallow==3.26.1
matplotlib==3.9.4
mpmath==1.3.0
mypy_extensions==1.1.0
narwhals @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_narwhals_1749208861/work
nest-asyncio==1.6.0
networkx==3.2.1
nltk==3.9.1
numpy @ file:///Users/runner/miniforge3/conda-bld/numpy_1732314388626/work/dist/numpy-2.0.2-cp39-cp39-macosx_11_0_arm64.whl#sha256=f5ae29dd6f735c6dcad61765cffcfb32661c04e515c759b21399c7a3a39430a4
olefile==0.47
ollama==0.5.1
omegaconf==2.3.0
onnx==1.18.0
onnxruntime==1.19.2
opencv-python==4.11.0.86
orjson==3.10.18
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1733203243479/work
pandas @ file:///Users/runner/miniforge3/conda-bld/pandas_1749099750763/work
parsing==2.0.4
pdf2image==1.17.0
pdfminer.six==20250506
pi_heif==0.22.0
pikepdf==9.8.1
pillow @ file:///Users/runner/miniforge3/conda-bld/pillow_1746646280087/work
pkgutil_resolve_name @ file:///home/conda/feedstock_root/build_artifacts/pkgutil-resolve-name_1733344503739/work
proto-plus==1.26.1
protobuf @ file:///Users/runner/miniforge3/conda-bld/protobuf_1741124632031/work/bazel-bin/python/dist/protobuf-5.29.3-cp39-abi3-macosx_11_0_arm64.whl#sha256=3b8a4cc9cbe1b96a58a05a21f7d8b4a3df6b04edc4ed11e4bfdd1d79fbe94ce1
psutil==7.0.0
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycocotools==2.0.10
pycparser @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_pycparser_1733195786/work
pydantic==2.11.5
pydantic_core==2.33.2
pydeck @ file:///home/conda/feedstock_root/build_artifacts/pydeck_1738346561094/work
pyparsing==3.2.3
pypdf==5.6.0
pypdfium2==4.30.1
PySocks @ file:///home/conda/feedstock_root/build_artifacts/pysocks_1733217236728/work
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1733215673016/work
python-iso639==2025.2.18
python-magic==0.4.27
python-multipart==0.0.20
python-oxmsg==0.0.2
pytz @ file:///home/conda/feedstock_root/build_artifacts/pytz_1742920838005/work
PyYAML @ file:///Users/runner/miniforge3/conda-bld/pyyaml_1737454732336/work
RapidFuzz==3.13.0
referencing @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_referencing_1737836872/work
regex==2024.11.6
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1733217035951/work
requests-toolbelt==1.0.0
rpds-py @ file:///Users/runner/miniforge3/conda-bld/bld/rattler-build_rpds-py_1747837782/work
rsa==4.9.1
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.13.1
sentence-transformers==4.1.0
six @ file:///home/conda/feedstock_root/build_artifacts/six_1733380938961/work
smmap @ file:///home/conda/feedstock_root/build_artifacts/smmap_1739781697784/work
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.41
streamlit @ file:///home/conda/feedstock_root/build_artifacts/streamlit_1747153904630/work
sympy==1.14.0
tenacity @ file:///home/conda/feedstock_root/build_artifacts/tenacity_1743640740204/work
threadpoolctl==3.6.0
timm==1.0.15
tokenizers==0.21.1
toml @ file:///home/conda/feedstock_root/build_artifacts/toml_1734091811753/work
torch==2.7.1
torchvision==0.22.1
tornado @ file:///Users/runner/miniforge3/conda-bld/tornado_1748003404111/work
tqdm==4.67.1
transformers==4.52.4
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_typing_extensions_1748959427/work
tzdata @ file:///home/conda/feedstock_root/build_artifacts/python-tzdata_1742745135198/work
unstructured==0.17.2
unstructured-client==0.36.0
unstructured-inference==1.0.5
unstructured.pytesseract==0.3.15
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1744323578849/work
utils==1.0.2
watchdog @ file:///Users/runner/miniforge3/conda-bld/watchdog_1730492905752/work
webencodings==0.5.1
wrapt==1.17.2
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1748277385522/work
zstandard==0.23.0
